+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @ai-sdk/langchain
 
+## 2.0.82
+
+### Patch Changes
+
+- ai@6.0.77
+
+## 2.0.81
+
+### Patch Changes
+
+- ai@6.0.76
+
+## 2.0.80
+
+### Patch Changes
+
+- Updated dependencies [7168375]
+  - ai@6.0.75
+
+## 2.0.79
+
+### Patch Changes
+
+- Updated dependencies [471009b]
+  - ai@6.0.74
+
+## 2.0.78
+
+### Patch Changes
+
+- ai@6.0.73
+
+## 2.0.77
+
+### Patch Changes
+
+- ai@6.0.72
+
+## 2.0.76
+
+### Patch Changes
+
+- ai@6.0.71
+
+## 2.0.75
+
+### Patch Changes
+
+- ai@6.0.70
+
+## 2.0.74
+
+### Patch Changes
+
+- Updated dependencies [d659305]
+  - ai@6.0.69
+
 ## 2.0.73
 
 ### Patch Changes